pyproject.toml
[tool.poetry]
name = "workout-tracker"
version = "0.1.0"
description = ""
authors = ["Damian Czajkowski <damian.czajkowski@mirumee.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.26.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
